app2/requirements.txt
flask
pdfminer.six
pandas
google.generativeai
flask-login
authlib